% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_linkage_functions.R
\name{cascade_link}
\alias{cascade_link}
\title{Title}
\usage{
cascade_link(
  tbl1,
  tbl2,
  link_instruction_tbl,
  name_tags = c("", ""),
  shrink_tbl1 = T,
  shrink_tbl2 = T
)
}
\arguments{
\item{tbl1}{First table to join}

\item{tbl2}{Second table to join}

\item{link_instruction_tbl}{A tibble of cascading join instructions}

\item{name_tags}{Optional name tags to suffix common variables}

\item{shrink_tbl1}{Whether to shink first table as it is iteratively joined. Default TRUE.}

\item{shrink_tbl2}{Whether to shink second table as it is iteratively joined. Default TRUE.}
}
\value{
A tibble of the cascade-joined tables
}
\description{
Title
}
\examples{
cascade_link(ed_tbl, mvc_tbl, ed_mvc_linkage_plan_tbl, name_tags = c("ed","mvc"))
}
