% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{get_shared_linkage_vars}
\alias{get_shared_linkage_vars}
\title{Get shared linkage variables as a list}
\usage{
get_shared_linkage_vars(tbl1, tbl2)
}
\arguments{
\item{tbl1}{Table 1}

\item{tbl2}{Table 2}
}
\value{
an intersection list of their common named linkage variables
}
\description{
Get shared linkage variables as a list
}
\examples{
get_shared_linkage_vars(this_tbl, that_tbl)
}
