% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_linkage_functions.R
\name{inner_join_rename}
\alias{inner_join_rename}
\title{Inner join two well formatted tables, renaming as we go}
\usage{
inner_join_rename(tbl1, tbl2, by_vars, tags = c("a", "b"))
}
\arguments{
\item{tbl1}{First table to join}

\item{tbl2}{Second table to join}

\item{by_vars}{Variables to join them on (TODO: make this optional, default to linkage vars)}

\item{tags}{Optional tags to add as suffix on common variables}
}
\value{
A tibble
}
\description{
Inner join two well formatted tables, renaming as we go
}
\examples{
link_tbl  = inner_join_rename(tbl1, tbl2, get_shared_linkage_vars(tbl1), tags = c("ED", "MVC"))
}
