% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{linkage_var_compare_tbl}
\alias{linkage_var_compare_tbl}
\title{Create a linkage variable comparison table}
\usage{
linkage_var_compare_tbl(tbl1, tbl2, tbl_names = c("tbl1", "tbl2"))
}
\arguments{
\item{tbl1}{First table}

\item{tbl2}{Second table}

\item{tbl_names}{A vector of table names for convenience}
}
\value{
A tibble
}
\description{
Create a linkage variable comparison table
}
\examples{
linkage_var_compare_tbl(this_tbl, that_tbl)
}
