% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{match_linkage_vars}
\alias{match_linkage_vars}
\title{Match linkage variables in a tidyselect pipeline}
\usage{
match_linkage_vars(match_regex = "^l_")
}
\arguments{
\item{match_regex}{An optional, alternate linkage variable regular expression to match on}
}
\value{
A matches() function in tidyselect format
}
\description{
Match linkage variables in a tidyselect pipeline
}
\examples{
this_tbl |> select(match_linkage_vars()) |> ...
}
