% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{select_linkage_vars}
\alias{select_linkage_vars}
\title{Select linkage variables by name in a table}
\usage{
select_linkage_vars(this_tbl)
}
\arguments{
\item{this_tbl}{A tibble}
}
\value{
A tibble
}
\description{
Select linkage variables by name in a table
}
\examples{
this_tbl |> select(match_linkage_vars()) |> ...
}
