% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{tag_link_vars}
\alias{tag_link_vars}
\title{Tag / rename linkage variables en masse}
\usage{
tag_link_vars(this_tbl, tag)
}
\arguments{
\item{this_tbl}{Table to tag linkage variables in}

\item{tag}{The tag to add as suffix}
}
\value{
A tibble
}
\description{
Tag / rename linkage variables en masse
}
\examples{
this_tbl = this_tbl |> tag_link_vars("EMS")
}
