% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_functions.R
\name{match_id_vars}
\alias{match_id_vars}
\title{Match id variables in a tidyselect pipeline}
\usage{
match_id_vars(match_regex = "^id_")
}
\arguments{
\item{match_regex}{An optional, alternate id variable regular expression to match on}
}
\value{
A matches() function in tidyselect format
}
\description{
Match id variables in a tidyselect pipeline
}
\examples{
this_tbl |> select(match_id_vars()) |> ...
}
